+++ src/tag/parser/compile/light.cpp
@@ -4,6 +4,19 @@
 
 namespace Invader::Parser {
     void Light::pre_compile(BuildWorkload &, std::size_t, std::size_t, std::size_t) {
+        // Clamp to 180 degrees
+        this->cutoff_angle = std::min(static_cast<double>(this->cutoff_angle), HALO_PI);
+        this->falloff_angle = std::min(static_cast<double>(this->falloff_angle), HALO_PI);
+        
+        // If cutoff angle is less than 0, set both to 180 degrees
+        if(this->cutoff_angle < 0.0) {
+            this->cutoff_angle = HALO_PI;
+            this->falloff_angle = HALO_PI;
+        }
+        
+        // Set falloff angle to cutoff angle if it's greater than it
+        this->falloff_angle = std::min(this->falloff_angle, this->cutoff_angle);
+        
         this->cos_cutoff_angle = std::cos(this->cutoff_angle);
         this->cos_falloff_angle = std::cos(this->falloff_angle);
         this->sin_cutoff_angle = std::sin(this->cutoff_angle);